+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-web/src/main/resources/META-INF/resources/js/ddm_form.js
@@ -700,15 +700,15 @@ AUI.add(
 								selectorInput.attr('disabled', instance.get('readOnly'));
 							}
 
-							var checkboxInput = container.one("input[type='checkbox']");
+							var checkboxInput = container.one('input[type="checkbox"]');
 
 							if (checkboxInput) {
 								checkboxInput.attr('disabled', instance.get('readOnly'));
 							}
 
-							var disableCheckboxInput = container.one("input[type='checkbox'][name$='disable']");
+							var disableCheckboxInput = container.one('input[type="checkbox"][name$="disable"]');
 
-							if (inputNode && disableCheckboxInput && disableCheckboxInput.get("checked")) {
+							if (inputNode && disableCheckboxInput && disableCheckboxInput.get('checked')) {
 								inputNode.attr('disabled', true);
 							}
 						}
@@ -2772,12 +2772,58 @@ AUI.add(
 								'render': instance._afterRenderTextHTMLField
 							}
 						);
+
+						var eventHandles = [
+							Liferay.on('inputLocalized:localeChanged', A.bind('_onLocaleChanged', instance))
+						];
+
+						instance._eventHandles = eventHandles;
+
+						instance._updateValues();
+					},
+
+					destructor: function() {
+						var instance = this;
+
+						(new A.EventHandle(instance._eventHandles)).detach();
 					},
 
 					getEditor: function() {
 						var instance = this;
 
 						return window[instance.getInputName() + 'Editor'];
+					},
+
+					getInputName: function() {
+						var instance = this;
+
+						var inputNode;
+
+						if (instance.get('localizable')) {
+							var fieldsNamespace = instance.get('fieldsNamespace');
+							var portletNamespace = instance.get('portletNamespace');
+
+							var prefix = [portletNamespace];
+
+							if (fieldsNamespace) {
+								prefix.push(fieldsNamespace);
+							}
+
+							inputNode = prefix.concat(
+								[
+									instance.get('name'),
+									'_',
+									INSTANCE_ID_PREFIX,
+									'_',
+									instance.get('instanceId')
+								]
+							).join('');
+						}
+						else {
+							inputNode = TextHTMLField.superclass.getInputName().apply(instance, arguments);
+						}
+
+						return inputNode;
 					},
 
 					getValue: function() {
@@ -2823,6 +2869,24 @@ AUI.add(
 						instance.get('container').toggle(!readOnly);
 					},
 
+					updateTranslationsDefaultValue: function() {
+						var instance = this;
+
+						var inputLocalized = Liferay.component(instance.getInputName());
+						var localizationMap = instance.get('localizationMap');
+
+						if (inputLocalized) {
+							inputLocalized.get('items').forEach(
+								function(item) {
+									localizationMap[item] = inputLocalized.getValue(item);
+								}
+							);
+						}
+						else {
+							TextHTMLField.superclass.updateTranslationsDefaultValue.apply(instance, arguments);
+						}
+					},
+
 					_afterRenderTextHTMLField: function() {
 						var instance = this;
 
@@ -2830,6 +2894,25 @@ AUI.add(
 
 						container.placeAfter(instance.readOnlyText);
 						container.placeAfter(instance.readOnlyLabel);
+					},
+
+					_onLocaleChanged: function(event) {
+						var instance = this;
+
+						var languageId = event.item.getAttribute('data-value');
+
+						instance.set('displayLocale', languageId);
+					},
+
+					_updateValues: function() {
+						var instance = this;
+
+						var inputLocalized = Liferay.component(instance.getInputName());
+						var localizationMap = instance.get('localizationMap');
+
+						for (var languageId in localizationMap) {
+							inputLocalized.updateInputLanguage(localizationMap[languageId], languageId);
+						}
 					}
 				}
 			}