+++ modules/apps/foundation/configuration-admin/configuration-admin-web/src/main/java/com/liferay/configuration/admin/web/internal/util/ConfigurationModelRetrieverImpl.java
@@ -109,6 +109,13 @@ public class ConfigurationModelRetrieverImpl
 		return null;
 	}
 
+	public ConfigurationCategory getConfigurationCategory(
+		String configurationCategoryKey) {
+
+		return _configurationCategoryServiceTrackerMap.getService(
+			configurationCategoryKey);
+	}
+
 	@Override
 	public ConfigurationCategoryMenuDisplay getConfigurationCategoryMenuDisplay(
 		String configurationCategory, String languageId) {
@@ -285,7 +292,7 @@ public class ConfigurationModelRetrieverImpl
 					ConfigurationCategory configurationCategory =
 						bundleContext.getService(serviceReference);
 
-					emitter.emit(configurationCategory.getKey());
+					emitter.emit(configurationCategory.getCategoryKey());
 				});
 	}
 
@@ -353,13 +360,6 @@ public class ConfigurationModelRetrieverImpl
 		}
 
 		return configurationCategories;
-	}
-
-	protected ConfigurationCategory getConfigurationCategory(
-		String configurationCategoryKey) {
-
-		return _configurationCategoryServiceTrackerMap.getService(
-			configurationCategoryKey);
 	}
 
 	protected ConfigurationModel getConfigurationModel(