+++ modules/sdk/gradle-plugins-defaults/src/main/java/com/liferay/gradle/plugins/defaults/internal/NodeDefaultsPlugin.java
@@ -21,15 +21,11 @@ import com.liferay.gradle.plugins.node.NodePlugin;
 import com.liferay.gradle.plugins.node.tasks.NpmInstallTask;
 import com.liferay.gradle.plugins.node.tasks.PublishNodeModuleTask;
 
-import java.io.File;
-
 import java.util.concurrent.Callable;
 
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.gradle.api.plugins.BasePlugin;
-import org.gradle.api.tasks.Delete;
 import org.gradle.api.tasks.TaskContainer;
 
 /**
@@ -41,7 +37,6 @@ public class NodeDefaultsPlugin extends BaseDefaultsPlugin<NodePlugin> {
 
 	@Override
 	protected void configureDefaults(Project project, NodePlugin nodePlugin) {
-		_configureTaskClean(project);
 		_configureTaskNpmInstall(project);
 		_configureTasksPublishNodeModule(project);
 	}
@@ -54,23 +49,11 @@ public class NodeDefaultsPlugin extends BaseDefaultsPlugin<NodePlugin> {
 	private NodeDefaultsPlugin() {
 	}
 
-	private void _configureTaskClean(Project project) {
-		boolean cleanNodeModules = Boolean.getBoolean("clean.node.modules");
-
-		if (cleanNodeModules) {
-			Delete delete = (Delete)GradleUtil.getTask(
-				project, BasePlugin.CLEAN_TASK_NAME);
-
-			delete.delete("node_modules");
-		}
-	}
-
 	private void _configureTaskNpmInstall(Project project) {
 		NpmInstallTask npmInstallTask = (NpmInstallTask)GradleUtil.getTask(
 			project, NodePlugin.NPM_INSTALL_TASK_NAME);
 
-		npmInstallTask.setNodeModulesDigestFile(
-			new File(npmInstallTask.getNodeModulesDir(), ".digest"));
+		npmInstallTask.setUseNpmCI(Boolean.TRUE);
 	}
 
 	private void _configureTaskPublishNodeModule(