+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/PortalRepositoryJob.java
@@ -32,7 +32,7 @@ public abstract class PortalRepositoryJob extends RepositoryJob {
 
 	@Override
 	public List<String> getBatchNames() {
-		String testBatchNames = portalTestProperies.getProperty(
+		String testBatchNames = portalTestProperties.getProperty(
 			"test.batch.names");
 
 		return getListFromString(testBatchNames);
@@ -40,10 +40,26 @@ public abstract class PortalRepositoryJob extends RepositoryJob {
 
 	@Override
 	public List<String> getDistTypes() {
-		String testBatchDistAppServers = portalTestProperies.getProperty(
+		String testBatchDistAppServers = portalTestProperties.getProperty(
 			"test.batch.dist.app.servers");
 
 		return getListFromString(testBatchDistAppServers);
+	}
+
+	public String getPoshiQuery(String testBatchName) {
+		String propertyName = JenkinsResultsParserUtil.combine(
+			"test.batch.run.property.query[", testBatchName, "]");
+
+		if (portalTestProperties.containsKey(propertyName)) {
+			String propertyValue = portalTestProperties.getProperty(
+				propertyName);
+
+			if (!propertyValue.isEmpty()) {
+				return propertyValue;
+			}
+		}
+
+		return null;
 	}
 
 	protected PortalRepositoryJob(String jobName) {
@@ -52,7 +68,7 @@ public abstract class PortalRepositoryJob extends RepositoryJob {
 		branchName = _getBranchName();
 		gitWorkingDirectory = _getGitWorkingDirectory();
 
-		portalTestProperies = getGitWorkingDirectoryProperties(
+		portalTestProperties = getGitWorkingDirectoryProperties(
 			"test.properties");
 	}
 
@@ -76,7 +92,7 @@ public abstract class PortalRepositoryJob extends RepositoryJob {
 		return list;
 	}
 
-	protected final Properties portalTestProperies;
+	protected final Properties portalTestProperties;
 
 	private String _getBranchName() {
 		Matcher matcher = _pattern.matcher(jobName);